view: optimize rendering, line-wise for same viewport.
save the old-screen as Vec<ScrLine> and for the next render()
use the cached screen-content and screen parameters to decide
whether to actually re-draw the line.

note that this algorithm is line oriented, and works well as
long as the view-port has not changed due to cursor motion.

for scroll-up and scroll-down a seperate optimization is needed.